On branch feature/model-ref  Issue No: #17
 Changes to be committed:
 Point 1: Create data using the .save() call should update the model instance with the record data. (_id attributed added and updated with the call to save())
 Point 2: find_one and find_all returns the model instance.
 Extension:- Converted the insert_one, insert_many, update_one, update_many, delete_one, delete_many to class method. As this needs to be called upon the model itself, not in one instance.
